requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
email-validator==2.1.0.post1
Flask-Migrate==4.0.7
gunicorn==21.2.0
